Restructuring of project and renaming
