+++ s/options.go
@@ -2,7 +2,9 @@ package s
 
 // This file all the constructor options for a SupervisorSpec
 
-import "github.com/capatazlib/go-capataz/c"
+import (
+	"github.com/capatazlib/go-capataz/c"
+)
 
 // WithOrder specifies the start/stop order of a supervisor's children
 func WithOrder(o Order) Opt {
